Template for data input
This is an excel template for the data input (the name of the file should be ¨input¨). The user must fill the columns ¨name¨(chemical name), ¨formula¨(neutral formula of the parent compound), M (neutral molecular weight), CAS (this should be in brackets, if not available it can be replaced by (na)) and RT (measured retention time in minutes). The first three rows are examples and can be deteleted by the user.